remove pkg/types/context
