+++ pkg/sessionctx/stmtctx/BUILD.bazel
@@ -12,7 +12,7 @@ go_library(
         "//pkg/parser/model",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
-        "//pkg/types/context",
+        "//pkg/types",
         "//pkg/util/disk",
         "//pkg/util/execdetails",
         "//pkg/util/intest",
@@ -47,7 +47,6 @@ go_test(
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
         "//pkg/types",
-        "//pkg/types/context",
         "//pkg/util/execdetails",
         "@com_github_pingcap_errors//:errors",
         "@com_github_stretchr_testify//require",
