+++ pkg/sessionctx/stmtctx/stmtctx.go
@@ -34,7 +34,7 @@ import (
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/pingcap/tidb/pkg/parser/terror"
-	typectx "github.com/pingcap/tidb/pkg/types/context"
+	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/disk"
 	"github.com/pingcap/tidb/pkg/util/execdetails"
 	"github.com/pingcap/tidb/pkg/util/intest"
@@ -156,7 +156,7 @@ type StatementContext struct {
 	_ constructor.Constructor `ctor:"NewStmtCtx,NewStmtCtxWithTimeZone,Reset"`
 
 	// 	typeCtx is used to indicate how to make the type conversation.
-	typeCtx typectx.Context
+	typeCtx types.Context
 
 	// Set the following variables before execution
 	StmtHints
@@ -425,7 +425,7 @@ type StatementContext struct {
 // NewStmtCtx creates a new statement context
 func NewStmtCtx() *StatementContext {
 	sc := &StatementContext{}
-	sc.typeCtx = typectx.NewContext(typectx.DefaultStmtFlags, time.UTC, sc.AppendWarning)
+	sc.typeCtx = types.NewContext(types.DefaultStmtFlags, time.UTC, sc.AppendWarning)
 	return sc
 }
 
@@ -433,14 +433,14 @@ func NewStmtCtx() *StatementContext {
 func NewStmtCtxWithTimeZone(tz *time.Location) *StatementContext {
 	intest.Assert(tz)
 	sc := &StatementContext{}
-	sc.typeCtx = typectx.NewContext(typectx.DefaultStmtFlags, tz, sc.AppendWarning)
+	sc.typeCtx = types.NewContext(types.DefaultStmtFlags, tz, sc.AppendWarning)
 	return sc
 }
 
 // Reset resets a statement context
 func (sc *StatementContext) Reset() {
 	*sc = StatementContext{
-		typeCtx: typectx.NewContext(typectx.DefaultStmtFlags, time.UTC, sc.AppendWarning),
+		typeCtx: types.NewContext(types.DefaultStmtFlags, time.UTC, sc.AppendWarning),
 	}
 }
 
@@ -456,17 +456,17 @@ func (sc *StatementContext) SetTimeZone(tz *time.Location) {
 }
 
 // TypeCtx returns the type context
-func (sc *StatementContext) TypeCtx() typectx.Context {
+func (sc *StatementContext) TypeCtx() types.Context {
 	return sc.typeCtx
 }
 
 // TypeFlags returns the type flags
-func (sc *StatementContext) TypeFlags() typectx.Flags {
+func (sc *StatementContext) TypeFlags() types.Flags {
 	return sc.typeCtx.Flags()
 }
 
 // SetTypeFlags sets the type flags
-func (sc *StatementContext) SetTypeFlags(flags typectx.Flags) {
+func (sc *StatementContext) SetTypeFlags(flags types.Flags) {
 	sc.typeCtx = sc.typeCtx.WithFlags(flags)
 }
 
@@ -1209,7 +1209,7 @@ func (sc *StatementContext) InitFromPBFlagAndTz(flags uint64, tz *time.Location)
 	sc.OverflowAsWarning = (flags & model.FlagOverflowAsWarning) > 0
 	sc.DividedByZeroAsWarning = (flags & model.FlagDividedByZeroAsWarning) > 0
 	sc.SetTimeZone(tz)
-	sc.SetTypeFlags(typectx.DefaultStmtFlags.
+	sc.SetTypeFlags(types.DefaultStmtFlags.
 		WithIgnoreTruncateErr((flags & model.FlagIgnoreTruncate) > 0).
 		WithTruncateAsWarning((flags & model.FlagTruncateAsWarning) > 0).
 		WithIgnoreZeroInDate((flags & model.FlagIgnoreZeroInDate) > 0).
@@ -1356,12 +1356,12 @@ func (sc *StatementContext) RecordedStatsLoadStatusCnt() (cnt int) {
 // If the statement context is nil, it'll return a newly created default type context.
 // **don't** use this function if you can make sure the `sc` is not nil. We should limit the usage of this function as
 // little as possible.
-func (sc *StatementContext) TypeCtxOrDefault() typectx.Context {
+func (sc *StatementContext) TypeCtxOrDefault() types.Context {
 	if sc != nil {
 		return sc.typeCtx
 	}
 
-	return typectx.DefaultStmtNoWarningContext
+	return types.DefaultStmtNoWarningContext
 }
 
 // UsedStatsInfoForTable records stats that are used during query and their information.
