+++ pkg/types/BUILD.bazel
@@ -36,6 +36,7 @@ go_library(
         "overflow.go",
         "set.go",
         "time.go",
+        "truncate.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/types",
     visibility = [
@@ -50,10 +51,10 @@ go_library(
         "//pkg/parser/opcode",
         "//pkg/parser/terror",
         "//pkg/parser/types",
-        "//pkg/types/context",
         "//pkg/util/collate",
         "//pkg/util/dbterror",
         "//pkg/util/hack",
+        "//pkg/util/intest",
         "//pkg/util/kvcache",
         "//pkg/util/logutil",
         "//pkg/util/mathutil",
@@ -74,6 +75,7 @@ go_test(
         "binary_literal_test.go",
         "compare_test.go",
         "const_test.go",
+        "context_test.go",
         "convert_test.go",
         "core_time_test.go",
         "datum_test.go",
@@ -103,7 +105,6 @@ go_test(
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
         "//pkg/testkit/testsetup",
-        "//pkg/types/context",
         "//pkg/util/collate",
         "//pkg/util/hack",
         "@com_github_pingcap_errors//:errors",
