+++ pkg/types/truncate.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package context
+package types
 
 import (
 	"github.com/pingcap/errors"
